requirements.txt
tensorflow==2.4.1
tensorflow_hub==0.11.0
Keras==2.4.3
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
opencv_python==4.5.1.48
streamlit==0.79.0
Pillow==8.1.2
